requirements.txt
python-telegram-bot
requests
tweepy
python-dotenv 
beautifulsoup4
yt-dlp